Add parallelization for playlist analytics
Spotify's analysis data for tracks is extremely detailed, but
unfortunately there is no batched API to fetch all this information
quickly. Thus, we resort to manually querying the API many times to
retrieve this information sequentially. This is usually very slow, since
the overhead from the repeated API calls adds up quickly. The obvious
solution is to parallelize this and perform all these transactions
concurrently.

This commit implements concurrent retrieval of analysis data for all
tracks within a playlist. Some quick testing shows significant speedups:

Pre:
    real    2m1.423s
    user    0m9.112s
    sys     0m3.228s

Post:
    real    0m26.849s
    user    0m11.894s
    sys     0m3.601s

All testing was done with a playlist of 328 items, likely much larger
than what we are going to encounter for day-to-day users, so hopefully
this current solution is fast enough going forward.